refactor: move apollo to its own file
